Block & Insert connection
Fleshing out how insert entities are drawn and how blocks are parsed and
stored.

Blocks hold references to entities are have coordinates relative to the
block versus regular entities that have coordinates relative to the
drawing. Insert entities supply the offsets and scale factors needed to
draw the entities contained within the block in the correct position. As
a result, a draw(InsertEntity) method has been added to the Entity
superclass to give blocks a method to call on their children.

This new draw overload adjusts the specific details of the subtype
entity, calls the standard draw method, and then adjusts them back to
unoffset values. This will leverage the draw code implementation for
each entity while allowing the block to modify its children based on the
specific insert entity that is being rendered.

- Removed Entity_Old code
- Temporarily fixed broken references in Hatch, Dimension to Entity_Old
- Adjusted Constructor for Schematic to include passing the Canvas.